chore: move nav_item to 1_component
